09_04/requirements.txt
pandas
numpy
opencv-python
Pillow
paddleocr
ultralytics
PyMuPDF
together
fitz    